+++ CHANGELOG.md
@@ -141,6 +141,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-2535 Dashboard statistics should only be computed on days having data
 - YLP-2536 Gain some space in patient statistics widget
 - YLP-2539 Update arrow icon in Device parameters change history
+- YLP-2554 Hide tooltip and values on TIR stat title when one bar is hovered
 
 ### Fixed
 - YLP-2145 Patient cannot see monitoring widgets when he's in a monitoring team and another non-monitoring team